refactor of new.vue and changed touch hold to double tap to delete album in chart
