+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "musicchartsite",
 	"description": "A website to create Music Charts",
-	"version": "0.8.0",
+	"version": "0.8.1",
 	"type": "module",
 	"author": {
 		"email": "musicChart@danielTurcich.com",
@@ -37,17 +37,17 @@
 		"@popperjs/core": "^2.11.6",
 		"dialog-polyfill": "^0.5.6",
 		"html2canvas": "^1.4.1",
-		"vue": "^3.2.40"
+		"vue": "^3.2.41"
 	},
 	"devDependencies": {
 		"@nabla/vite-plugin-eslint": "^1.4.1",
 		"@types/node": "18.8.0",
-		"@typescript-eslint/eslint-plugin": "^5.40.0",
-		"@typescript-eslint/parser": "^5.40.0",
-		"@vitejs/plugin-vue": "^3.1.2",
+		"@typescript-eslint/eslint-plugin": "^5.41.0",
+		"@typescript-eslint/parser": "^5.41.0",
+		"@vitejs/plugin-vue": "^3.2.0",
 		"@vue/eslint-config-typescript": "^11.0.2",
 		"autoprefixer": "^10.4.12",
-		"eslint": "^8.25.0",
+		"eslint": "^8.26.0",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-prettier": "^4.2.1",
 		"eslint-plugin-vue": "^9.6.0",
@@ -55,10 +55,10 @@
 		"postcss": "^8.4.18",
 		"prettier": "^2.7.1",
 		"prettier-plugin-tailwindcss": "^0.1.13",
-		"tailwindcss": "^3.1.8",
+		"tailwindcss": "^3.2.1",
 		"typescript": "^4.8.4",
-		"vite": "^3.1.7",
+		"vite": "^3.2.0",
 		"vue-eslint-parser": "^9.1.0",
-		"vue-tsc": "^1.0.6"
+		"vue-tsc": "^1.0.9"
 	}
 }
