+++ src/components/SidebarComponents/SelectionComponents/New.vue
@@ -20,8 +20,8 @@ import {
 	SiteOptionsKey
 } from '#shared/storage'
 
-import Back from '#assets/back.svg'
 import Dialog from '#core/Dialog.vue'
+import Presets from './NewComponents/Presets.vue'
 
 const props = defineProps<{
 	selectedChartTitle: string
@@ -38,8 +38,6 @@ const presetAdd = ref(false)
 const chartNameInput = ref(props.selectedChartTitle)
 const chartInput = ref<HTMLInputElement>()
 const createPreset = ref(false)
-const newPresetName = ref<string>()
-const newPresetList = ref<string>()
 const editPresets = ref(false)
 
 function onNewChart() {
@@ -82,30 +80,6 @@ function saveCurrentChart() {
 		return setStoredChart(props.selectedChartTitle, props.selectedChart!)
 	}
 	throw new Error('Error: Cannot save current Chart!')
-}
-
-function closeCreatePreset() {
-	createPreset.value = false
-	newPresetName.value = undefined
-	newPresetList.value = undefined
-}
-
-// Make sure that this is not called unless its a valid number seperated list
-function newPreset() {
-	const massagedPresetList = newPresetList.value!.split(',').map(Number)
-
-	GlobalSiteOptions.presets?.push({
-		default: false,
-		rowSizes: massagedPresetList,
-		presetName: newPresetName.value
-	})
-	closeCreatePreset()
-}
-
-function deletePreset(preset: ChartPreset, index: number) {
-	if (window.confirm(`Delete preset '${preset.presetName}'?'`)) {
-		GlobalSiteOptions.presets?.splice(index, 1)
-	}
 }
 
 onMounted(() => {
@@ -153,107 +127,14 @@ onMounted(() => {
 				</div>
 
 				<template v-else>
-					<template v-if="!createPreset && !editPresets">
-						<div class="tw-flex-center gap-1">
-							<button
-								v-for="(preset, index) in GlobalSiteOptions.presets"
-								:key="`preset-${index}`"
-								class="tw-button"
-								@click="newChart('Preset', preset)"
-							>
-								{{ preset.presetName }}
-							</button>
-						</div>
-
-						<div class="flex gap-1 mt-2">
-							<button class="tw-button" @click="createPreset = true">
-								New preset
-							</button>
-							<button
-								v-if="
-									GlobalSiteOptions.presets?.some((preset) => !preset.default)
-								"
-								class="tw-button"
-								@click="editPresets = true"
-							>
-								Edit presets
-							</button>
-						</div>
-						<button
-							class="tw-button tw-close-button"
-							@click="presetAdd = false"
-						>
-							<img :src="Back" width="25" height="25" />
-						</button>
-					</template>
-
-					<template v-if="createPreset">
-						<h1 class="text-lg text-green-400">Create a new preset</h1>
-
-						<ol class="mx-4 text-left list-decimal">
-							<li>The input should be a list of numbers</li>
-							<li>Seperate numbers by commas</li>
-							<li>You can't include a 0 in the list</li>
-							<li>These can be really weird, go for it</li>
-						</ol>
-
-						<div class="flex gap-1">
-							<p class="text-green-400">Example 1 -</p>
-							5, 5, 6, 6, 10, 10
-						</div>
-						<div class="flex gap-1">
-							<p class="text-green-400">Example 2 -</p>
-							15,3,8,5
-						</div>
-
-						<input
-							v-model="newPresetName"
-							class="tw-input"
-							placeholder="Preset Name"
-						/>
-						<input
-							v-model="newPresetList"
-							class="tw-input"
-							placeholder="New preset input"
-						/>
-
-						<div class="flex gap-1">
-							<button class="tw-button" @click="newPreset">
-								Create preset
-							</button>
-							<button
-								class="tw-button tw-close-button"
-								@click="closeCreatePreset"
-							>
-								<img :src="Back" width="25" height="25" />
-							</button>
-						</div>
-					</template>
-
-					<template v-if="editPresets">
-						Edit preset name or delete
-						<template
-							v-for="(preset, index) in GlobalSiteOptions.presets"
-							:key="`edit-preset-${index}`"
-						>
-							<div v-if="!preset.default" class="flex gap-2">
-								<input
-									v-model="GlobalSiteOptions.presets![index].presetName"
-									class="tw-input"
-								/>
-								<button class="tw-button" @click="deletePreset(preset, index)">
-									Delete
-								</button>
-							</div>
-						</template>
-
-						<button
-							class="tw-button tw-close-button"
-							@click="editPresets = false"
-						>
-							<img :src="Back" width="25" height="25" />
-						</button>
-					</template>
+					<Presets
+						:create-preset="createPreset"
+						:edit-presets="editPresets"
+						@update-create-preset="(value) => (createPreset = value)"
+						@update-edit-presets="(value) => (editPresets = value)"
+						@update-preset-add="(value) => (presetAdd = value)"
+						@new-chart="newChart"
+					/>
 				</template>
 			</template>
 		</template>
